Add files via upload
This file codes for a game called Multiplication Quiz. It's a math game containing simple multiplication questions. The game contains 3 different difficulty levels as well as an infinite mode. The game is also able to keep track of the player's score and tell them which answer was wrong along with the right answer.