Add files via upload
This is a game/quiz called the Barbie Name Generator. The player is able to answer questions and in the end they'll receive a character name from Barbie: Life in the Dreamhouse. Different interests will align with different characters and depending on what the player's interests are, they will receive whichever Barbie character they're most similar to.